+++ ecomarkets/src/main/java/ecomarkets/rs/PartnerResource.java
@@ -43,6 +43,4 @@ public class PartnerResource {
         .build();
     }
 
-    
-
 }
