+++ ecomarkets/src/main/java/ecomarkets/rs/basket/BasketResource.java
@@ -3,6 +3,8 @@ package ecomarkets.rs.basket;
 import ecomarkets.domain.core.basket.Basket;
 import ecomarkets.domain.core.basket.event.BasketEvent;
 import ecomarkets.domain.core.product.Product;
+import ecomarkets.rs.basket.form.BasketItemForm;
+import ecomarkets.rs.basket.form.CreateBasketForm;
 import io.quarkus.panache.common.Sort;
 import jakarta.enterprise.event.Event;
 import jakarta.inject.Inject;
