+++ ecomarkets/src/main/java/ecomarkets/rs/product/CategoryResource.java
@@ -58,10 +58,4 @@ public class CategoryResource {
         }
     }
 
-
-
-
-
-    
-
 }
